Added static mesh collision and visibility bbox support. 1) Fixed phantom step on (TR3: house.tr2); 2) Fixed state change information loading: no freezing animations in some Meta2tr levels; 3) Fixed trimesh generation function (resource.cpp) - no crashes in degenerate bbox cases; 4) Fixed controls: this time used native SDL2 keycodes for ctrl and shift:     bind(actions.key_action, 0x400000E0, 0x400000E4); -- ctrl case;     bind(actions.key_walk, 0x400000E1, 0x400000E5);   -- shift case;

+++ engine.h
@@ -56,8 +56,8 @@ typedef struct engine_control_mapper_s
 
     int8_t   joy_axis_map[AXIS_LASTINDEX];      // Axis array for action mapper.
 
-    int16_t  action_map[ACT_LASTINDEX];         // Actions array for action mapper.
-    int16_t  action_alt[ACT_LASTINDEX];         // Alternate actions array for alternate controls.
+    int32_t  action_map[ACT_LASTINDEX];         // Actions array for action mapper.
+    int32_t  action_alt[ACT_LASTINDEX];         // Alternate actions array for alternate controls.
 
 }engine_control_mapper_t, *engine_control_mapper_p;
